// content/videos.generated.ts
export type VideoCategory = {
  slug: string;
  label: { es: string; en: string };
  count: number;
};

export const videoCategories: VideoCategory[] = [
  {
    "slug": "documentales",
    "label": {
      "es": "Documentales",
      "en": "Documentaries"
    },
    "count": 10
  },
  {
    "slug": "musica",
    "label": {
      "es": "Música",
      "en": "Music"
    },
    "count": 4
  }
];